perf: Use a linked list for vram slot allocation
Tiles are allocated into a structure of arrays double linked list,
instead of a table. The linked list includes a reference count to
handle duplicates.

While new method this is more complicated then a simple table, it is
twice as fast.

This occurs because:
    1) Only the allocated slots are checked for duplicate tilesets
    2) The search code fits in the three registers the 65816 has.

This update has also changed the size of the tileset slot reference
counter from sint16 to uint8.

Profiling
=========
Profiling the code is done for the Unit Test code.

It saves an average of 2.25 scanlines per allocation.

Old:
    80859a MetaSprite__UploadFixedTileset 614694 cycles, called 109 times

New:
    8085fe MetaSprite__UploadFixedTileset 280684 cycles, called 109 times